Return size, subs, ... as doubles to match tensor toolbox

+++ matlab/matlab_src/@sptensor_gt/size.m
@@ -21,7 +21,7 @@ function m = size(t,idx)
 
 
 if exist('idx','var')
-    m = t.size(idx);
+    m = double(t.size(idx));
 else
-    m = t.size;
+    m = double(t.size);
 end
